Oldi

+++ inc/ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: qclubfoo <qclubfoo@student.42.fr>          +#+  +:+       +#+        */
+/*   By: djeanna <djeanna@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/24 08:39:30 by djeanna           #+#    #+#             */
-/*   Updated: 2019/04/28 16:31:48 by qclubfoo         ###   ########.fr       */
+/*   Updated: 2019/04/29 17:39:50 by djeanna          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,13 +39,13 @@ int					ft_printf(char *str, ...);
 
 size_t				ft_percent(va_list ap, char **str, t_param list);
 
-void				ft_putnbr(intmax_t n);
+void				ft_putnbr(long long n);
 void				ft_putchar(char c);
 void				ft_putstr(char const *s);
 int					ft_putnstr(char *str, int size);
 size_t				ft_strlen(const char *s);
 int					ft_print_symb(int c, int size);
-int					ft_num_dig(intmax_t n, int base);
+int					ft_num_dig(long long n, int base);
 
 int					ft_c(t_param list, int c);
 int					ft_s(t_param list, char *s);
@@ -58,7 +58,7 @@ int					ft_prcnt(t_param list);
 char				*ft_strchr(const char *s, int c);
 
 int					ft_atoi(const char *str);
-char				*ft_itoa(intmax_t n);
-char				*ft_itoa_base(intmax_t n, int base);
+char				*ft_itoa(long long n);
+char				*ft_itoa_base(long long n, int base);
 
 #endif
