+++ src/ft_itoa_base.c
@@ -6,13 +6,14 @@
 /*   By: djeanna <djeanna@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/08 10:19:47 by djeanna           #+#    #+#             */
-/*   Updated: 2019/04/27 13:16:20 by djeanna          ###   ########.fr       */
+/*   Updated: 2019/04/29 17:39:06 by djeanna          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/ft_printf.h"
 
-static void		turn_into_char(char *res, intmax_t n, int base, int num_d)
+static void		turn_into_char(char *res, long long n,
+								int base, int num_d)
 {
 	char	*buf;
 	int		sign;
@@ -29,7 +30,7 @@ static void		turn_into_char(char *res, intmax_t n, int base, int num_d)
 	}
 }
 
-char			*ft_itoa_base(intmax_t n, int base)
+char			*ft_itoa_base(long long n, int base)
 {
 	char	*res;
 	int		num_d;
